refactor: decreasing the degree of penalty for inconsistent payment

+++ rinha-2025-rafael/Domain/SummaryDetails.cs
@@ -1,4 +1,4 @@
 ﻿namespace rinha_2025_rafael.Domain
 {
-    public record SummaryDetails(long TotalRequests, double TotalAmount);
+    public record SummaryDetails(long TotalRequests, decimal TotalAmount);
 }
